+++ kite/src/main/kotlin/com/tang/kite/annotation/logical/LogicalDelete.kt
@@ -1,0 +1,9 @@
+package com.tang.kite.annotation.logical
+
+/**
+ * @author Tang
+ */
+@Target(AnnotationTarget.FIELD)
+@Retention(AnnotationRetention.RUNTIME)
+@MustBeDocumented
+annotation class LogicalDelete()