+++ kite/src/main/kotlin/com/tang/kite/config/KiteConfig.kt
@@ -3,6 +3,7 @@ package com.tang.kite.config
 import com.tang.kite.annotation.fill.CreateTime
 import com.tang.kite.annotation.fill.UpdateTime
 import com.tang.kite.config.defaults.DefaultSelectiveStrategy
+import com.tang.kite.config.logical.LogicalDeleteConfig
 import com.tang.kite.config.table.TableConfig
 import com.tang.kite.enumeration.SqlType
 import com.tang.kite.handler.fill.FillHandler
@@ -69,5 +70,11 @@ object KiteConfig {
      */
     @JvmStatic
     val table = TableConfig
+
+    /**
+     * Logical delete properties for logical delete configuration.
+     */
+    @JvmStatic
+    val logicalDelete = LogicalDeleteConfig
 
 }