+++ kite/src/main/kotlin/com/tang/kite/config/logical/LogicalDeleteValue.kt
@@ -1,0 +1,12 @@
+package com.tang.kite.config.logical
+
+/**
+ * @author Tang
+ */
+data class LogicalDeleteValue(
+
+    val normalValue: Any,
+
+    val deletedValue: Any
+
+)