+++ kite/src/test/kotlin/com/tang/kite/session/mapper/LogicalAccountMapper.kt
@@ -1,0 +1,9 @@
+package com.tang.kite.session.mapper
+
+import com.tang.kite.mapper.BaseMapper
+import com.tang.kite.session.entity.LogicalAccount
+
+/**
+ * @author Tang
+ */
+interface LogicalAccountMapper : BaseMapper<LogicalAccount>